# new_server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "new_server API",
    "version": "1.0.0",
    "description": "Machine-readable API contract for the Express backend in new_server."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyAuthScheme": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative format: `Authorization: ApiKey <apiKey>`"
      },
      "inboundBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for worker ingress"
      }
    },
    "schemas": {
      "RequestId": {
        "type": "string"
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "requestId"],
        "properties": {
          "error": { "type": "string" },
          "details": {},
          "requestId": { "$ref": "#/components/schemas/RequestId" }
        }
      },
...
      "EmailSender": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "name": { "type": "string", "nullable": true },
              "address": { "type": "string", "nullable": true }
            }
          },
          {
            "type": "null"
          }
        ]
      },
      "Email": {
        "type": "object",
        "required": ["id", "to", "localPart", "domain", "envelopeFrom", "from", "subject", "text", "html", "workerName", "sourceDomain", "messageId", "receivedAt", "createdAt", "hasRawMime", "rawMimeSize", "groupCount"],
        "properties": {
          "id": { "type": "integer" },
          "to": { "type": "string" },
          "localPart": { "type": "string" },
          "domain": { "type": "string" },
          "envelopeFrom": { "type": "string", "nullable": true },
          "from": { "$ref": "#/components/schemas/EmailSender" },
          "subject": { "type": "string" },
          "text": { "type": "string" },
          "html": { "type": "string" },
          "workerName": { "type": "string" },
          "sourceDomain": { "type": "string" },
          "messageId": { "type": "string" },
          "receivedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "hasRawMime": { "type": "boolean" },
          "rawMimeSize": { "type": "integer" },
          "groupCount": { "type": "integer" },
          "rawMime": { "type": "string", "nullable": true }
        }
      },
      "EmailRegister": {
        "type": "object",
        "required": ["id", "ownerUserId", "owner", "emailAddress", "localPart", "domain", "emailCount", "latestReceivedAt", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer" },
          "ownerUserId": { "type": "integer" },
          "owner": { "$ref": "#/components/schemas/GroupOwner" },
          "emailAddress": { "type": "string" },
          "localPart": { "type": "string" },
          "domain": { "type": "string" },
          "emailCount": { "type": "integer" },
          "latestReceivedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "GroupOwner": {
        "oneOf": [
          {
            "type": "object",
            "required": ["id", "username", "displayName"],
            "properties": {
              "id": { "type": "integer" },
              "username": { "type": "string" },
              "displayName": { "type": "string" }
            }
          },
          {
            "type": "null"
          }
        ]
      },
      "Group": {
        "type": "object",
        "required": ["id", "ownerUserId", "owner", "name", "color", "description", "emailCount", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer" },
          "ownerUserId": { "type": "integer" },
          "owner": { "$ref": "#/components/schemas/GroupOwner" },
          "name": { "type": "string" },
          "color": { "type": "string" },
          "description": { "type": "string" },
          "emailCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "GroupEmail": {
        "allOf": [
          { "$ref": "#/components/schemas/Email" },
          {
            "type": "object",
            "required": ["groupPosition", "groupAddedAt", "groupAddedByUserId"],
            "properties": {
              "groupPosition": { "type": "integer" },
              "groupAddedAt": { "type": "string", "format": "date-time" },
              "groupAddedByUserId": { "type": "integer", "nullable": true }
            }
          }
        ]
      },
...
      "BatchEmailRequest": {
        "type": "object",
        "required": ["emailIds"],
        "properties": {
          "emailIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 200,
            "items": {
              "oneOf": [{ "type": "integer" }, { "type": "string" }]
            }
          },
          "includeRawMime": { "type": "boolean" }
        }
      },
      "EmailRegisterCreateRequest": {
        "type": "object",
        "required": ["emailAddress"],
        "properties": {
          "emailAddress": { "type": "string" }
        }
      },
...
      "GroupEmailBatchRequest": {
        "type": "object",
        "required": ["emailIds"],
        "properties": {
          "emailIds": {
            "type": "array",
            "maxItems": 200,
            "items": {
              "oneOf": [{ "type": "integer" }, { "type": "string" }]
            }
          },
          "includeRawMime": { "type": "boolean" }
        }
      },
      "SimpleSuccessResponse": {
        "type": "object",
        "required": ["success", "requestId"],
        "properties": {
          "success": { "type": "boolean" },
          "requestId": { "$ref": "#/components/schemas/RequestId" }
        }
      }
    },
    "responses": {
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/email-registers": {
      "get": {
        "tags": ["EmailRegisters"],
        "summary": "List mailbox registrations for the current user",
        "description": "Normal users only see their own registrations. Global admins may pass ownerUserId to inspect another user.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "query", "name": "ownerUserId", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Registration list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "registrations": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/EmailRegister" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["EmailRegisters"],
        "summary": "Register a mailbox for monitoring",
        "description": "Mailbox must still be covered by the caller's permission scope. Each mailbox can only be registered by one user globally. Re-registering the same mailbox by the same user is idempotent.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmailRegisterCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mailbox registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "registration": { "$ref": "#/components/schemas/EmailRegister" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/v1/email-registers/{registrationId}": {
      "delete": {
        "tags": ["EmailRegisters"],
        "summary": "Delete a mailbox registration",
        "description": "Owners may delete their own registration. Global admins may delete any registration. Removing a registration also removes matching emails from the owner's groups.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "registrationId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Registration deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SimpleSuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/emails": {
      "get": {
        "tags": ["Emails"],
        "summary": "List emails",
        "description": "Global admins can list all emails without filters. Normal users only see emails whose mailbox is both permitted and registered in email_registers for that user. domain/address filters narrow the result set.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "query", "name": "domain", "schema": { "type": "string" } },
          { "in": "query", "name": "address", "schema": { "type": "string" } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 50, "maximum": 200 } },
          { "in": "query", "name": "offset", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "Email list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": { "type": "integer" },
                    "count": { "type": "integer" },
                    "emails": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Email" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad query or missing scope filters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/emails/batch": {
      "post": {
        "tags": ["Emails"],
        "summary": "Batch fetch emails by ID with partial success",
        "description": "Partial success response. deniedIds includes emails outside permission scope and, for normal users, emails whose mailbox is not registered. Caller should handle missingIds and deniedIds separately.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchEmailRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "emails": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Email" }
                    },
                    "missingIds": {
                      "type": "array",
                      "items": { "type": "integer" }
                    },
                    "deniedIds": {
                      "type": "array",
                      "items": { "type": "integer" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/emails/{id}": {
      "get": {
        "tags": ["Emails"],
        "summary": "Get email by ID",
        "description": "Normal users can only read emails whose mailbox is registered for them and still covered by permission. Global admins can read any email.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } },
          { "in": "query", "name": "includeRawMime", "schema": { "type": "string", "enum": ["1"] } }
        ],
        "responses": {
          "200": {
            "description": "Email details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "$ref": "#/components/schemas/Email" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Emails"],
        "summary": "Delete email by ID",
        "description": "Requires operator-or-higher mailbox permission. Normal users must also have the mailbox registered for themselves.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Email deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SimpleSuccessResponse" }
              }
            }
          }
        }
      }
    },
...
    "/v1/groups/{groupId}/emails": {
      "get": {
        "tags": ["Groups"],
        "summary": "List group emails",
        "description": "Group visibility is evaluated against the group owner's mailbox registrations and permissions. If any saved email id is now missing or denied, the backend removes it from the group and returns 409 so the client can reload.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 100, "maximum": 200 } },
          { "in": "query", "name": "offset", "schema": { "type": "integer", "default": 0 } },
          { "in": "query", "name": "includeRawMime", "schema": { "type": "string", "enum": ["1"] } }
        ],
        "responses": {
          "200": {
            "description": "Group emails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "group": { "$ref": "#/components/schemas/Group" },
                    "total": { "type": "integer" },
                    "count": { "type": "integer" },
                    "emails": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/GroupEmail" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "post": {
        "tags": ["Groups"],
        "summary": "Append emails to group",
        "description": "A group is permanently owned by one user. Every emailId added must resolve to a mailbox registered by that owner, otherwise the request is rejected.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GroupEmailBatchRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Emails appended to group",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "group": { "$ref": "#/components/schemas/Group" },
                    "total": { "type": "integer" },
                    "count": { "type": "integer" },
                    "emails": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/GroupEmail" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Groups"],
        "summary": "Replace group emails",
        "description": "Replaces the entire group membership. Every emailId must still be visible to the group owner through both permission scope and mailbox registration.",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GroupEmailBatchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group emails replaced",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "group": { "$ref": "#/components/schemas/Group" },
                    "total": { "type": "integer" },
                    "count": { "type": "integer" },
                    "emails": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/GroupEmail" }
                    },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/groups/{groupId}/emails/{emailId}": {
      "delete": {
        "tags": ["Groups"],
        "summary": "Remove one email from group",
        "security": [{ "bearerAuth": [] }, { "apiKeyHeader": [] }, { "apiKeyAuthScheme": [] }],
        "parameters": [
          { "in": "path", "name": "groupId", "required": true, "schema": { "type": "integer" } },
          { "in": "path", "name": "emailId", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Email removed from group",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SimpleSuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/inbound/email": {
      "post": {
        "tags": ["Inbound"],
        "summary": "Ingest raw MIME email from worker",
        "security": [{ "inboundBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "message/rfc822": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "parameters": [
          { "in": "header", "name": "X-Email-Envelope-To", "required": false, "schema": { "type": "string" } },
          { "in": "header", "name": "X-Email-Envelope-From", "required": false, "schema": { "type": "string" } },
          { "in": "header", "name": "X-Email-Worker-Name", "required": false, "schema": { "type": "string" } },
          { "in": "header", "name": "X-Email-Received-At", "required": false, "schema": { "type": "string" } },
          { "in": "header", "name": "X-Email-Domain", "required": false, "schema": { "type": "string" } },
          { "in": "header", "name": "X-Email-Message-Id", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "202": {
            "description": "Email accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "id": { "type": "integer" },
                    "envelopeTo": { "type": "string" },
                    "domain": { "type": "string" },
                    "receivedAt": { "type": "string", "format": "date-time" },
                    "requestId": { "$ref": "#/components/schemas/RequestId" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}